requirements.txt
litellm
mysql-connector-python
python-dotenv
flask